Removing upgrade reports and removing old files.
