--- research/bert/create_pretraining_data_upgrade_report.txt
@@ -1,147 +1,0 @@
---------------------------------------------------------------------------------
-Processing file 'create_pretraining_data.py'
- outputting to 'create_pretraining_data-upgraded.py'
---------------------------------------------------------------------------------
-
-'create_pretraining_data.py' Line 97
---------------------------------------------------------------------------------
-
-Renamed function 'tf.python_io.TFRecordWriter' to 'tf.io.TFRecordWriter'
-
-    Old:     writers.append(tf.python_io.TFRecordWriter(output_file))
-                            ~~~~~~~~~~~~~~~~~~~~~~~~~~~               
-    New:     writers.append(tf.io.TFRecordWriter(output_file))
-                            ~~~~~~~~~~~~~~~~~~~~               
-
-'create_pretraining_data.py' Line 162
---------------------------------------------------------------------------------
-
-Renamed function 'tf.logging.info' to 'tf.compat.v1.logging.info'
-
-    Old:   tf.logging.info("Wrote %d total instances", total_written)
-           ~~~~~~~~~~~~~~~                                            
-    New:   tf.compat.v1.logging.info("Wrote %d total instances", total_written)
-           ~~~~~~~~~~~~~~~~~~~~~~~~~                                            
-
-'create_pretraining_data.py' Line 419
---------------------------------------------------------------------------------
-
-Renamed function 'tf.logging.info' to 'tf.compat.v1.logging.info'
-
-    Old:   tf.logging.info("*** Reading from input files ***")
-           ~~~~~~~~~~~~~~~                                     
-    New:   tf.compat.v1.logging.info("*** Reading from input files ***")
-           ~~~~~~~~~~~~~~~~~~~~~~~~~                                     
-
-'create_pretraining_data.py' Line 421
---------------------------------------------------------------------------------
-
-Renamed function 'tf.logging.info' to 'tf.compat.v1.logging.info'
-
-    Old:     tf.logging.info("  %s", input_file)
-             ~~~~~~~~~~~~~~~                     
-    New:     tf.compat.v1.logging.info("  %s", input_file)
-             ~~~~~~~~~~~~~~~~~~~~~~~~~                     
-
-'create_pretraining_data.py' Line 417
---------------------------------------------------------------------------------
-
-Renamed function 'tf.gfile.Glob' to 'tf.io.gfile.glob'
-
-    Old:     input_files.extend(tf.gfile.Glob(input_pattern))
-                                ~~~~~~~~~~~~~                 
-    New:     input_files.extend(tf.io.gfile.glob(input_pattern))
-                                ~~~~~~~~~~~~~~~~                 
-
-'create_pretraining_data.py' Line 188
---------------------------------------------------------------------------------
-
-Renamed function 'tf.gfile.GFile' to 'tf.io.gfile.GFile'
-
-    Old:     with tf.gfile.GFile(input_file, "r") as reader:
-                  ~~~~~~~~~~~~~~                             
-    New:     with tf.io.gfile.GFile(input_file, "r") as reader:
-                  ~~~~~~~~~~~~~~~~~                             
-
-'create_pretraining_data.py' Line 430
---------------------------------------------------------------------------------
-
-Renamed function 'tf.logging.info' to 'tf.compat.v1.logging.info'
-
-    Old:   tf.logging.info("*** Writing to output files ***")
-           ~~~~~~~~~~~~~~~                                    
-    New:   tf.compat.v1.logging.info("*** Writing to output files ***")
-           ~~~~~~~~~~~~~~~~~~~~~~~~~                                    
-
-'create_pretraining_data.py' Line 432
---------------------------------------------------------------------------------
-
-Renamed function 'tf.logging.info' to 'tf.compat.v1.logging.info'
-
-    Old:     tf.logging.info("  %s", output_file)
-             ~~~~~~~~~~~~~~~                      
-    New:     tf.compat.v1.logging.info("  %s", output_file)
-             ~~~~~~~~~~~~~~~~~~~~~~~~~                      
-
-'create_pretraining_data.py' Line 145
---------------------------------------------------------------------------------
-
-Renamed function 'tf.logging.info' to 'tf.compat.v1.logging.info'
-
-    Old:       tf.logging.info("*** Example ***")
-               ~~~~~~~~~~~~~~~                    
-    New:       tf.compat.v1.logging.info("*** Example ***")
-               ~~~~~~~~~~~~~~~~~~~~~~~~~                    
-
-'create_pretraining_data.py' Line 146
---------------------------------------------------------------------------------
-
-Renamed function 'tf.logging.info' to 'tf.compat.v1.logging.info'
-
-    Old:       tf.logging.info("tokens: %s" % " ".join(
-               ~~~~~~~~~~~~~~~                          
-    New:       tf.compat.v1.logging.info("tokens: %s" % " ".join(
-               ~~~~~~~~~~~~~~~~~~~~~~~~~                          
-
-'create_pretraining_data.py' Line 26
---------------------------------------------------------------------------------
-
-tf.flags has been removed, please use the argparse or absl module if you need command line parsing.
-
-    Old: flags = tf.flags
-                 ~~~~~~~~ 
-    New: flags = tf.flags
-                 ~~~~~~~~ 
-
-'create_pretraining_data.py' Line 156
---------------------------------------------------------------------------------
-
-Renamed function 'tf.logging.info' to 'tf.compat.v1.logging.info'
-
-    Old:         tf.logging.info(
-                 ~~~~~~~~~~~~~~~  
-    New:         tf.compat.v1.logging.info(
-                 ~~~~~~~~~~~~~~~~~~~~~~~~~  
-
-'create_pretraining_data.py' Line 410
---------------------------------------------------------------------------------
-
-Renamed function 'tf.logging.set_verbosity' to 'tf.compat.v1.logging.set_verbosity'
-Renamed function 'tf.logging.INFO' to 'tf.compat.v1.logging.INFO'
-
-    Old:   tf.logging.set_verbosity(tf.logging.INFO)
-           ~~~~~~~~~~~~~~~~~~~~~~~~ ~~~~~~~~~~~~~~~  
-    New:   tf.compat.v1.logging.set_verbosity(tf.compat.v1.logging.INFO)
-           ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ~~~~~~~~~~~~~~~~~~~~~~~~~  
-
-'create_pretraining_data.py' Line 442
---------------------------------------------------------------------------------
-
-Renamed function 'tf.app.run' to 'tf.compat.v1.app.run'
-
-    Old:   tf.app.run()
-           ~~~~~~~~~~   
-    New:   tf.compat.v1.app.run()
-           ~~~~~~~~~~~~~~~~~~~~   
-
-
