--- research/bert/optimization_upgrade_report.txt
@@ -1,67 +1,0 @@
---------------------------------------------------------------------------------
-Processing file 'optimization.py'
- outputting to 'optimization-upgraded.py'
---------------------------------------------------------------------------------
-
-'optimization.py' Line 32
---------------------------------------------------------------------------------
-
-WARNING: tf.train.polynomial_decay has been changed to return a callable instead of a tensor when graph building, but its functionality remains unchanged during eager execution (returns a callable like before). The converter cannot detect and fix this reliably, so this usage has been converted to compat.v1 (even though it may already be correct).
-
-
-    Old:   learning_rate = tf.train.polynomial_decay(
-                           ~~~~~~~~~~~~~~~~~~~~~~~~~  
-    New:   learning_rate = tf.train.polynomial_decay(
-                           ~~~~~~~~~~~~~~~~~~~~~~~~~  
-
-'optimization.py' Line 27
---------------------------------------------------------------------------------
-
-Renamed function 'tf.train.get_or_create_global_step' to 'tf.compat.v1.train.get_or_create_global_step'
-
-    Old:   global_step = tf.train.get_or_create_global_step()
-                         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~   
-    New:   global_step = tf.compat.v1.train.get_or_create_global_step()
-                         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~   
-
-'optimization.py' Line 70
---------------------------------------------------------------------------------
-
-Renamed function 'tf.trainable_variables' to 'tf.compat.v1.trainable_variables'
-
-    Old:   tvars = tf.trainable_variables()
-                   ~~~~~~~~~~~~~~~~~~~~~~   
-    New:   tvars = tf.compat.v1.trainable_variables()
-                   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~   
-
-'optimization.py' Line 117
---------------------------------------------------------------------------------
-
-Renamed function 'tf.get_variable' to 'tf.compat.v1.get_variable'
-
-    Old:       m = tf.get_variable(
-                   ~~~~~~~~~~~~~~~  
-    New:       m = tf.compat.v1.get_variable(
-                   ~~~~~~~~~~~~~~~~~~~~~~~~~  
-
-'optimization.py' Line 87
---------------------------------------------------------------------------------
-
-Renamed function 'tf.train.Optimizer' to 'tf.compat.v1.train.Optimizer'
-
-    Old: class AdamWeightDecayOptimizer(tf.train.Optimizer):
-                                        ~~~~~~~~~~~~~~~~~~   
-    New: class AdamWeightDecayOptimizer(tf.compat.v1.train.Optimizer):
-                                        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~   
-
-'optimization.py' Line 123
---------------------------------------------------------------------------------
-
-Renamed function 'tf.get_variable' to 'tf.compat.v1.get_variable'
-
-    Old:       v = tf.get_variable(
-                   ~~~~~~~~~~~~~~~  
-    New:       v = tf.compat.v1.get_variable(
-                   ~~~~~~~~~~~~~~~~~~~~~~~~~  
-
-
