--- research/bert/tokenization_test_upgrade_report.txt
@@ -1,6 +1,0 @@
---------------------------------------------------------------------------------
-Processing file 'tokenization_test.py'
- outputting to 'tokenization_test-upgraded.py'
---------------------------------------------------------------------------------
-
-
