--- research/bert/tokenization_upgrade_report.txt
@@ -1,16 +1,0 @@
---------------------------------------------------------------------------------
-Processing file 'tokenization.py'
- outputting to 'tokenization-upgraded.py'
---------------------------------------------------------------------------------
-
-'tokenization.py' Line 125
---------------------------------------------------------------------------------
-
-Renamed function 'tf.gfile.GFile' to 'tf.io.gfile.GFile'
-
-    Old:   with tf.gfile.GFile(vocab_file, "r") as reader:
-                ~~~~~~~~~~~~~~                             
-    New:   with tf.io.gfile.GFile(vocab_file, "r") as reader:
-                ~~~~~~~~~~~~~~~~~                             
-
-
